requirements/base.txt
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1
drf-spectacular==0.27.2
celery==5.3.6
redis==5.0.4
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-json-logger==2.0.7
Pillow==10.3.0
whitenoise==6.6.0
gunicorn==22.0.0
